Refactor: annotation package 정리, Profile 응답 로직 수정
