+++ src/main/java/com/knud4/an/account/dto/ProfileListResponse.java
@@ -8,10 +8,10 @@ import java.util.stream.Collectors;
 
 @Data
 public class ProfileListResponse {
-    private final List<ProfileDTO> profiles;
+    private final List<ProfileDTO> list;
 
-    public ProfileListResponse(List<Profile> profiles) {
-        this.profiles = profiles.stream().map(p -> new ProfileDTO(p.getId(), p.getName(), p.getAge(), p.getGender()))
+    public ProfileListResponse(List<Profile> list) {
+        this.list = list.stream().map(p -> new ProfileDTO(p.getId(), p.getName(), p.getAge(), p.getGender()))
                 .collect(Collectors.toList());
     }
 }
