+++ src/main/java/com/knud4/an/community/controller/CommunityController.java
@@ -8,8 +8,8 @@ import com.knud4.an.community.dto.CommunityListDTO;
 import com.knud4.an.community.dto.CreateCommunityForm;
 import com.knud4.an.community.entity.Category;
 import com.knud4.an.community.service.CommunityService;
-import com.knud4.an.interceptor.AccountRequired;
-import com.knud4.an.interceptor.ProfileRequired;
+import com.knud4.an.annotation.AccountRequired;
+import com.knud4.an.annotation.ProfileRequired;
 import com.knud4.an.utils.api.ApiUtil;
 import com.knud4.an.utils.api.ApiUtil.*;
 import io.swagger.v3.oas.annotations.Operation;
