+++ src/main/java/com/knud4/an/notice/controller/NoticeController.java
@@ -3,8 +3,8 @@ package com.knud4.an.notice.controller;
 import com.knud4.an.account.entity.Profile;
 import com.knud4.an.account.service.AccountService;
 import com.knud4.an.board.Range;
-import com.knud4.an.interceptor.AccountRequired;
-import com.knud4.an.interceptor.ProfileRequired;
+import com.knud4.an.annotation.AccountRequired;
+import com.knud4.an.annotation.ProfileRequired;
 import com.knud4.an.notice.dto.CreateNoticeForm;
 import com.knud4.an.notice.dto.NoticeDTO;
 import com.knud4.an.notice.dto.NoticeListDTO;
